dashboard/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
influxdb-client==1.45.0
websockets==13.1
python-dotenv==1.0.1
jinja2==3.1.4